Make Query optional and error-throwing
